dashboard/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from .models import Notifications

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['username', 'id']


class NotificationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notifications
        fields = '__all__'